Refactored out the tweet message building into a TweetBuilder interface
